requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
django-environ>=0.11.0
Pillow>=10.0.0
psycopg2-binary>=2.9.9
requests>=2.31.0
gunicorn>=21.2.0